Added support for float

+++ src/main.rs
@@ -1,12 +1,12 @@
 use std::io;
 
-fn calculator(val_1: i32, val_2: i32, opr: &str) -> Result<i32, &str> {
+fn calculator(val_1: f32, val_2: f32, opr: &str) -> Result<f32, &str> {
     match opr {
         "+" => Ok(val_1 + val_2),
         "-" => Ok(val_1 - val_2),
         "*" => Ok(val_1 * val_2),
         "/" => {
-            if val_2 == 0 {
+            if val_2 == 0.0 {
                 return Err("Cannot divide by zero!");
             }
         Ok(val_1 / val_2)
@@ -24,13 +24,13 @@ fn main() {
     io::stdin()
         .read_line(&mut input_1)
         .expect("Failed to read the first value.");
-    let input_1: i32 = input_1.trim().parse().expect("Please enter a number!");
+    let input_1: f32 = input_1.trim().parse().expect("Please enter a number!");
     
     eprint!("Enter the second value: ");
     io::stdin()
         .read_line(&mut input_2)
         .expect("Failed to read the second value.");
-    let input_2: i32 = input_2.trim().parse().expect("Please enter a number!");
+    let input_2: f32 = input_2.trim().parse().expect("Please enter a number!");
 
     eprint!("Enter the operator: ");
     io::stdin()
